refactor(usb): split xhci.rs into a module directory
Summary:
  - Move the 2842-line driver into usb/xhci/{mod,ring,context,init,
    commands,slots,transfers,enumerate,configure,bulk}.rs with zero
    logic changes: method bodies land verbatim, moved methods gain
    pub(super) for cross-file calls within the driver subtree, and each
    file gets exact (compiler-verified) imports since bare names never
    resolve to parent scope
  - TrbRing fields go pub(super) for ring-state polling; UsbSlot was
    already all-pub; test-only xhci-crate layout imports move into the
    tests module; external paths (XhciController, XhciError, MAX_SLOTS,
    do_*_transfer) are unchanged

Tests:
  - cargo fmt --check
  - cargo clippy --workspace --release -- -D warnings; ui and fb-log
    features; aarch64 and riscv64 targets
  - cargo test --workspace (xhci layout tests moved with the code)
  - ./crabefi build for x86-64, aarch64 (virt/sbsa), riscv64
  - xtask tests, coreboot feature script
